Move to upstart for Aurora service control
Using rc.local was always a bad idea and now we are seeing the fallout.
When a Playa Mesos environment is provisioned with accp.bash, the
rc.local is overwritten by the slave provisioner and the master
provisioner since they both live on the same host.

Migrate the scheduler and thermos observer startup scripts to upstart
for long-term sanity.